long 类型比较

+++ src/main/java/com/wangzhixuan/controller/UserController.java
@@ -135,7 +135,7 @@ public class UserController extends BaseController {
     @ResponseBody
     public Object edit(UserVo userVo) {
         User user = userService.selectByLoginName(userVo.getLoginName());
-        if (user != null && user.getId() != userVo.getId()) {
+        if (user != null && user.getId().intValue() != userVo.getId().intValue()) {
             return renderError("用户名已存在!");
         }
         if (StringUtils.isNotBlank(userVo.getPassword())) {
